Lots more functionality work, mostly across block entity components, fixing misc. bugs and architecture work
- EMI tag translation, ignore entries
- Book recipe fixups
- Block, item tag work
- Fixed ingot pile model render crash
- Rework for most block entity renderers, and how they access textures/models, split to be independent client-side initialization, and generally extensible
- Cleanup, rework for client-size inventory tooltips, work on making block entity item components functional at all
- Cleanup for `Helpers` methods, compress some methods, inline `openScreen`
- Fix food not being eaten
- Fix decaying in creative tabs in search
- Lots of migrations to use `DeferredRegister.getEntries()` where we can rather than mod ID name filtering

+++ src/main/java/net/dries007/tfc/world/chunkdata/RegionChunkDataGenerator.java
@@ -118,11 +118,11 @@ public record RegionChunkDataGenerator(
 
         // Distance within the grid of this chunk - so a value between [0, 1] representing the top left of this chunk
         // The interpolator will add 16 / <grid width> to obtain the other side of this chunk, and interpolate from the bounding boxes of the grid points.
-        final double deltaX = exactGridX - gridX;
-        final double deltaZ = exactGridZ - gridZ;
+        final float deltaX = (float) (exactGridX - gridX);
+        final float deltaZ = (float) (exactGridZ - gridZ);
 
         // The width of the chunk, in grid coordinates
-        final double dG = Units.blockToGridExact(16);
+        final float dG = (float) Units.blockToGridExact(16);
 
         // The base rainfall and temperature layers, scaled down to chunk resolution
         LerpFloatLayer rainfallLayer = rainfallGridLayer.scaled(deltaX, deltaZ, dG);
